chat_managment/srcs/requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.5
django-channels-presence-4.0==1.1.2
django-cors-headers==4.6.0
django-htmx==1.21.0
django_redis==5.4.0
# Suppression de django-rest-framework==0.1.0 qui provoque une erreur
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-django==4.9.0
python-decouple==3.8
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.8.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uvicorn==0.27.1
uvicorn[standard]
websockets==12.0
websocket-client==1.8.0
whitenoise==6.6.0
zope.interface==7.2
watchdog